Enable group movable for all elements.
